move row to suitable place, and improve comments.
